raddrinfo.c: suppress a warning
* ext/socket/raddrinfo.c (parse_numeric_port): explict cast to
  suppress a warning for precision losing conversion.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@50894 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/socket/raddrinfo.c
@@ -175,7 +175,7 @@ parse_numeric_port(const char *service, int *portp)
     if (0x10000 <= u)
         return 0;
 
-    *portp = u;
+    *portp = (int)u;
 
     return 1;
 }
